[#95] Refactor: ChatRes dto 분리
